backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.5.2               # via kombu
asgiref==3.2.7            # via django
billiard==3.6.3.0         # via celery
celery==4.4.2             # via -r requirements.in
click==7.1.1              # via pip-tools
django==3.0.4             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
fasttext==0.9.1           # via -r requirements.in
kombu==4.6.8              # via celery
numpy==1.18.3             # via fasttext
pip-tools==4.5.1          # via -r requirements.in
pybind11==2.5.0           # via fasttext
python-decouple==3.3      # via -r requirements.in
pytz==2019.3              # via celery, django
redis==3.4.1              # via -r requirements.in
six==1.14.0               # via pip-tools
sqlparse==0.3.1           # via django
vine==1.3.0               # via amqp, celery

# The following packages are considered to be unsafe in a requirements file:
# setuptools
